requirements.txt
Django==1.8.6
pytz==2015.7
wheel==0.24.0
django-widget-tweaks==1.4.1
django-taggit==0.17.1
Markdown==2.6.2

django-haystack==2.4.0
pysolr==3.3.2